+++ tests/test_tokens_v2.py
@@ -188,6 +188,25 @@ class TestTokensV2(CaptureLogMixin, CreateUserMixin, TestCase):
         self.assertEqual(user, self.user)
         self.assertLogsContain("Valid token for user john in default scope")
 
+    # Test token invalidation on email change
+
+    def test_valid_token_after_email_change(self):
+        token = create_token(self.user)
+        self.user.email = "email@example.com"
+        self.user.save()
+        user = parse_token(token, self.get_user)
+        self.assertEqual(user, self.user)
+        self.assertLogsContain("Valid token for user john in default scope")
+
+    @override_settings(SESAME_INVALIDATE_ON_EMAIL_CHANGE=True)
+    def test_invalid_token_after_email_change(self):
+        token = create_token(self.user)
+        self.user.email = "email@example.com"
+        self.user.save()
+        user = parse_token(token, self.get_user)
+        self.assertIsNone(user)
+        self.assertLogsContain("Invalid token for user john in default scope")
+
     # Test scoped tokens
 
     def test_valid_scoped_token_in_scope(self):